feat(ssg): :art: decoupling markdown.rs
