+++ src/modules/mod.rs
@@ -22,6 +22,9 @@ pub mod keywords;
 /// The `manifest` module generates the manifest file.
 pub mod manifest;
 
+/// The `markdown` module contains functions for converting Markdown to HTML.
+pub mod markdown;
+
 /// The `metatags` module contains functions for generating meta tags.
 pub mod metatags;
